add reviews router
